+++ src/client/client.ts
@@ -1,7 +1,8 @@
 // This file runs in the browser.
 
-// injected by serverPluginHmr when served
-declare const __PORT__: number
+// injected by serverPluginClient when served
+declare const __HMR_PROTOCOL__: string
+declare const __HMR_HOST__: string
 declare const __MODE__: string
 declare const __DEFINES__: Record<string, any>
 ;(window as any).process = (window as any).process || {}
@@ -29,9 +30,10 @@ console.log('[vite] connecting...')
 
 declare var __VUE_HMR_RUNTIME__: HMRRuntime
 
-const socketProtocol = location.protocol === 'https:' ? 'wss' : 'ws'
-const socketUrl = `${socketProtocol}://${location.hostname}:${__PORT__}`
-const socket = new WebSocket(socketUrl, 'vite-hmr')
+// use server configuration, then fallback to inference
+const socketProtocol =
+  __HMR_PROTOCOL__ || (location.protocol === 'https:' ? 'wss' : 'ws')
+const socket = new WebSocket(`${socketProtocol}://${__HMR_HOST__}`, 'vite-hmr')
 
 function warnFailedFetch(err: Error, path: string | string[]) {
   if (!err.message.match('fetch')) {
@@ -225,11 +227,9 @@ async function updateModule(
 ) {
   const mod = hotModulesMap.get(id)
   if (!mod) {
-    console.error(
-      `[vite] got js update notification for "${id}" but no client callback ` +
-        `was registered. Something is wrong.`
-    )
-    console.error(hotModulesMap)
+    // In a code-spliting project,
+    // it is common that the hot-updating module is not loaded yet.
+    // https://github.com/vitejs/vite/issues/721
     return
   }