+++ src/node/resolver.ts
@@ -15,8 +15,9 @@ import {
 } from './server/serverPluginModuleResolve'
 import { resolveOptimizedCacheDir } from './optimizer'
 import { clientPublicPath } from './server/serverPluginClient'
+import { isCSSRequest } from './utils/cssUtils'
+import { isStaticAsset } from './utils/pathUtils'
 import chalk from 'chalk'
-import { isAsset } from './optimizer/pluginAssets'
 
 const debug = require('debug')('vite:resolve')
 const isWin = require('os').platform() === 'win32'
@@ -38,6 +39,7 @@ export interface InternalResolver {
     relativePublicPath: string
   ): { pathname: string; query: string }
   isPublicRequest(publicPath: string): boolean
+  isAssetRequest(filePath: string): boolean
 }
 
 export const supportedExts = ['.mjs', '.js', '.ts', '.jsx', '.tsx', '.json']
@@ -98,8 +100,9 @@ const resolveFilePathPostfix = (filePath: string): string | undefined => {
         postfix = ext
         break
       }
-      if (isFile(path.join(cleanPath, '/index' + ext))) {
-        postfix = '/index' + ext
+      const defaultFilePath = `/index${ext}`
+      if (isFile(path.join(cleanPath, defaultFilePath))) {
+        postfix = defaultFilePath
         break
       }
     }
@@ -118,7 +121,8 @@ const isDir = (p: string) => fs.existsSync(p) && fs.statSync(p).isDirectory()
 export function createResolver(
   root: string,
   resolvers: Resolver[] = [],
-  userAlias: Record<string, string> = {}
+  userAlias: Record<string, string> = {},
+  assetsInclude?: (file: string) => boolean
 ): InternalResolver {
   resolvers = [...resolvers]
   const literalAlias: Record<string, string> = {}
@@ -155,9 +159,9 @@ export function createResolver(
     }
   }
 
-  resolvers.forEach((r) => {
-    if (r.alias && typeof r.alias === 'object') {
-      resolveAlias(r.alias)
+  resolvers.forEach(({ alias }) => {
+    if (alias && typeof alias === 'object') {
+      resolveAlias(alias)
     }
   })
   resolveAlias(userAlias)
@@ -167,6 +171,7 @@ export function createResolver(
 
   const resolver: InternalResolver = {
     requestToFile(publicPath) {
+      publicPath = decodeURIComponent(publicPath)
       if (requestToFileCache.has(publicPath)) {
         return requestToFileCache.get(publicPath)!
       }
@@ -280,9 +285,8 @@ export function createResolver(
       if (aliased) {
         return aliased
       }
-      for (const r of resolvers) {
-        aliased =
-          r.alias && typeof r.alias === 'function' ? r.alias(id) : undefined
+      for (const { alias } of resolvers) {
+        aliased = alias && typeof alias === 'function' ? alias(id) : undefined
         if (aliased) {
           return aliased
         }
@@ -325,6 +329,12 @@ export function createResolver(
       return resolver
         .requestToFile(publicPath)
         .startsWith(path.resolve(root, 'public'))
+    },
+
+    isAssetRequest(filePath: string) {
+      return (
+        (assetsInclude && assetsInclude(filePath)) || isStaticAsset(filePath)
+      )
     }
   }
 
@@ -364,8 +374,7 @@ export function resolveBareModuleRequest(
     if (!pkgInfo.entry) {
       console.error(
         chalk.yellow(
-          `[vite] dependency ${id} does not have default entry defined in ` +
-            `package.json.`
+          `[vite] dependency ${id} does not have default entry defined in package.json.`
         )
       )
     } else {
@@ -387,7 +396,7 @@ export function resolveBareModuleRequest(
           // redirect it the optimized copy.
           return resolveBareModuleRequest(root, depId, importer, resolver)
         }
-        if (!isAsset(id)) {
+        if (!isCSSRequest(id) && !resolver.isAssetRequest(id)) {
           // warn against deep imports to optimized dep
           console.error(
             chalk.yellow(
@@ -509,7 +518,7 @@ export function resolveNodeModule(
 
     // resolve object browser field in package.json
     // https://github.com/defunctzombie/package-browser-field-spec
-    const browserField = pkg.browser
+    const { browser: browserField } = pkg
     if (entryPoint && browserField && typeof browserField === 'object') {
       entryPoint = mapWithBrowserField(entryPoint, browserField)
     }
@@ -582,9 +591,9 @@ function mapWithBrowserField(
   map: Record<string, string>
 ) {
   const normalized = normalize(relativePathInPkgDir)
-  const foundEntry = Object.entries(map).find(([from]) => {
-    return normalize(from) === normalized
-  })
+  const foundEntry = Object.entries(map).find(
+    ([from]) => normalize(from) === normalized
+  )
   if (!foundEntry) {
     return normalized
   }