+++ src/node/server/serverPluginProxy.ts
@@ -1,7 +1,10 @@
 import { ServerPlugin } from '.'
 import { URL } from 'url'
+import { IKoaProxiesOptions } from 'koa-proxies'
 
-export const proxyPlugin: ServerPlugin = ({ app, config }) => {
+export type ProxiesOptions = IKoaProxiesOptions & { ws: boolean }
+
+export const proxyPlugin: ServerPlugin = ({ app, config, server }) => {
   if (!config.proxy) {
     return
   }
@@ -12,8 +15,9 @@ export const proxyPlugin: ServerPlugin = ({ app, config }) => {
   Object.keys(options).forEach((path) => {
     let opts = options[path]
     if (typeof opts === 'string') {
-      opts = { target: opts }
+      opts = { target: opts } as ProxiesOptions
     }
+    if (opts.ws) return
     opts.logs = (ctx, target) => {
       debug(
         `${ctx.req.method} ${(ctx.req as any).oldPath} proxy to -> ${new URL(
@@ -23,5 +27,16 @@ export const proxyPlugin: ServerPlugin = ({ app, config }) => {
       )
     }
     app.use(proxy(path, opts))
+  })
+
+  server.on('upgrade', (req, socket, head) => {
+    if (req.headers['sec-websocket-protocol'] !== 'vite-hmr') {
+      for (const path in options) {
+        let opts = options[path]
+        if (typeof opts === 'object' && opts.ws) {
+          proxy.proxy.ws(req, socket, head, opts)
+        }
+      }
+    }
   })
 }