+++ test/test.js
@@ -3,7 +3,6 @@ const path = require('path')
 const execa = require('execa')
 const puppeteer = require('puppeteer')
 const moment = require('moment')
-const { EPERM } = require('constants')
 
 jest.setTimeout(100000)
 
@@ -801,6 +800,16 @@ describe('vite', () => {
       // should be inside the async chunk
       expect(code).toMatch(colorToMatch)
     })
+
+    test('build manifest', async () => {
+      const manifest = JSON.parse(
+        await fs.readFile(path.join(tempDir, 'dist/_assets/manifest.json'))
+      )
+      const indexPath = manifest['index.js']
+      expect(
+        await fs.stat(path.join(tempDir, `dist/_assets/${indexPath}`))
+      ).toBeTruthy()
+    })
   })
 
   describe('dev', () => {
@@ -835,7 +844,7 @@ describe('vite', () => {
     declareTests(false)
 
     test('hmr (index.html full-reload)', async () => {
-      expect(await getText('title')).toMatch('Vite App')
+      expect(await getText('title')).toMatch('Vite Playground')
       // hmr
       const reload = page.waitForNavigation({
         waitUntil: 'domcontentloaded'
@@ -844,12 +853,12 @@ describe('vite', () => {
         content.replace('Vite App', 'Vite App Test')
       )
       await reload
-      await expectByPolling(() => getText('title'), 'Vite App Test')
+      await expectByPolling(() => getText('title'), 'Vite Playground Test')
     })
 
     test('hmr (html full-reload)', async () => {
       await page.goto('http://localhost:3000/test.html')
-      expect(await getText('title')).toMatch('Vite App')
+      expect(await getText('title')).toMatch('Vite Playground')
       // hmr
       const reload = page.waitForNavigation({
         waitUntil: 'domcontentloaded'
@@ -858,7 +867,7 @@ describe('vite', () => {
         content.replace('Vite App', 'Vite App Test')
       )
       await reload
-      await expectByPolling(() => getText('title'), 'Vite App Test')
+      await expectByPolling(() => getText('title'), 'Vite Playground Test')
     })
 
     // Assert that all edited files are reflected on page reload